NeighborlyAPI/requirements.txt
azure-functions==1.2.1
pymongo==3.10.1
azure-eventgrid==4.9.0
